Update marquee.ino
It seems to me that the library for ledmatrix allows brightness values from 0 to 15 (it is even commented in line 178 of marquee.ino). However the web interface limits the value to 1-15. I do not see the reason for the limitation and would love to use the clock in my bedroom but it's simply too bright at value 1. This is my first proposition on github so I hope it is okay and sorry if I made any mistakes.